% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn_forecast.R
\name{wnn_fit}
\alias{wnn_fit}
\title{Train Weighted Nearest Neighbours model}
\usage{
wnn_fit(series, lag = 96, k = 10)
}
\arguments{
\item{series}{Numeric time series}

\item{lag}{Embedding dimension}

\item{k}{Number of nearest neighbours}
}
\description{
Train Weighted Nearest Neighbours model
}
