% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn_forecast.R
\name{wnn_forecast}
\alias{wnn_forecast}
\title{Forecast using WNN model}
\usage{
wnn_forecast(model, h = 96)
}
\arguments{
\item{model}{Model created by wnn_fit}

\item{h}{Forecast horizon}
}
\description{
Forecast using WNN model
}
